huge refactors
added many many grid tool helpers to abstract core grid math into condensed functions. changed a lot of naming for clarity. Next move is to extract out all the 'os' styled features and focus in on core terminal/textmode stuff because i got caught in the weeds of styled rects when the OG intention was textmode